fix imports and wrong device
