+++ attacks/decision/decision_black_box_attack.py
@@ -1,3 +1,6 @@
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
 '''
 
 This file is copied from the following source:
@@ -25,10 +28,6 @@ basic structure for main:
 """
 Implements the base class for decision-based black-box attacks
 """
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
 import numpy as np
 import torch
 from torch import Tensor as t
