+++ attacks/decision/sign_flip_attack.py
@@ -1,3 +1,6 @@
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
 '''
 
 This file is modified based on the following source:
@@ -27,10 +30,6 @@ basic structure for main:
 """
 Implements Sign Flip
 """
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
 import numpy as np
 import torch
 from torch import Tensor as t
