+++ attacks/score/bandit_attack.py
@@ -1,3 +1,6 @@
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
 '''
 
 This file is copied from the following source:
@@ -24,10 +27,6 @@ basic structure for main:
 """
 Implements Bandit Attack from Ilyas et al. 2018
 """
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
 import numpy as np
 import torch
