+++ attacks/score/simple_attack.py
@@ -1,3 +1,6 @@
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
 '''
 
 This file is copied from the following source:
@@ -16,31 +19,27 @@ url={https://openreview.net/forum?id=SygW0TEFwH}
 
 basic structure for main:
     1. config args and prior setup
-    2. define functions that insert perturbation and return the estimated gradient
+    2. define functions that insert perturbation using simple attack
     3. return results
-
+    
 '''
 
 """
-Implements SignHunter
+Implements the simple black-box attack
 """
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
 import numpy as np
-import torch 
+import torch
 
 from attacks.score.score_black_box_attack import ScoreBlackBoxAttack
-from utils.compute import lp_step, sign, norm
+from utils.compute import lp_step, sign
 
 
-class SignAttack(ScoreBlackBoxAttack):
+class SimpleAttack(ScoreBlackBoxAttack):
     """
-    SignHunter
+    Simple Black-Box Attack
     """
 
-    def __init__(self, max_loss_queries, epsilon, p, fd_eta, lb, ub, batch_size, name):
+    def __init__(self, max_loss_queries, epsilon, p, lb, ub, delta, batch_size, name):
         """
         :param max_loss_queries: maximum number of calls allowed to loss oracle per data pt
         :param epsilon: radius of lp-ball of perturbation
@@ -55,100 +54,86 @@ class SignAttack(ScoreBlackBoxAttack):
                          p=p,
                          lb=lb,
                          ub=ub,
-                         batch_size= batch_size,
-                         name="Sign")
-
-
-        self.fd_eta = fd_eta
-        self.best_est_deriv = None
-        self.xo_t = None
-        self.sgn_t = None
-        self.h = 0
+                         batch_size = batch_size,
+                         name = "SimBA")
+                         
+        #self.xo_t = None
+        self.delta = delta
+        self.perm = None
+        self.best_loss = None
         self.i = 0
 
     def _perturb(self, xs_t, loss_fct):
         _shape = list(xs_t.shape)
         dim = np.prod(_shape[1:])
-        # additional queries at the start
+        b_sz = _shape[0]
         add_queries = 0
         if self.is_new_batch:
-            self.xo_t = xs_t.clone()
-            self.h = 0
+            #self.xo_t = xs_t.clone()
             self.i = 0
-        if self.i == 0 and self.h == 0:
-            self.sgn_t = sign(torch.ones(_shape[0], dim))
-            fxs_t = lp_step(self.xo_t, self.sgn_t.view(_shape), self.epsilon, self.p)
-            bxs_t = self.xo_t  
-            est_deriv = (loss_fct(fxs_t) - loss_fct(bxs_t)) / self.epsilon
-            self.best_est_deriv = est_deriv
-            add_queries = 3  # because of bxs_t and the 2 evaluations in the i=0, h=0, case.
-        chunk_len = np.ceil(dim / (2 ** self.h)).astype(int)
-        istart = self.i * chunk_len
-        iend = min(dim, (self.i + 1) * chunk_len)
-        self.sgn_t[:, istart:iend] *= - 1.
-        fxs_t = lp_step(self.xo_t, self.sgn_t.view(_shape), self.epsilon, self.p)
-        bxs_t = self.xo_t
-        est_deriv = (loss_fct(fxs_t) - loss_fct(bxs_t)) / self.epsilon
-        
-        ### sign here
-        self.sgn_t[[i for i, val in enumerate(est_deriv < self.best_est_deriv) if val], istart: iend] *= -1.
-        
-
-        self.best_est_deriv = (est_deriv >= self.best_est_deriv) * est_deriv + (
-                est_deriv < self.best_est_deriv) * self.best_est_deriv
-        # perform the step
-        new_xs = lp_step(self.xo_t, self.sgn_t.view(_shape), self.epsilon, self.p)
-        # update i and h for next iteration
+            self.perm = torch.rand(b_sz, dim).argsort(dim=1)
+        if self.i == 0:
+            #self.sgn_t = sign(torch.ones(_shape[0], dim))
+            #fxs_t = lp_step(self.xo_t, self.sgn_t.view(_shape), self.epsilon, self.p)
+            #bxs_t = self.xo_t
+            loss = loss_fct(xs_t)
+            self.best_loss = loss
+            add_queries = 1
+        diff = torch.zeros(b_sz, dim)
+        # % if iterations are greater than dim
+        idx = self.perm[:, self.i % dim]
+        diff = diff.scatter(1, idx.unsqueeze(1), 1)
+        new_xs = xs_t.clone().contiguous().view(b_sz,-1)
+        # left attempt
+        left_xs = lp_step(xs_t, diff.view_as(xs_t), self.delta, self.p)
+        left_loss = loss_fct(left_xs)
+        replace_flag = torch.tensor((left_loss > self.best_loss).astype(np.float32)).unsqueeze(1)
+        #print(replace_flag.shape)
+        self.best_loss = replace_flag.squeeze(1) * left_loss + (1 - replace_flag.squeeze(1)) * self.best_loss
+        new_xs = replace_flag * left_xs.contiguous().view(b_sz,-1) + (1. - replace_flag) * new_xs
+        # right attempt
+        right_xs = lp_step(xs_t, diff.view_as(xs_t), - self.delta, self.p)
+        right_loss = loss_fct(right_xs)
+        # replace only those that have greater right loss and was not replaced
+        # in the left attempt
+        replace_flag = torch.tensor((right_loss > self.best_loss).astype(np.float32)).unsqueeze(1) * (1 - replace_flag)
+        #print(replace_flag.shape)
+        self.best_loss = replace_flag.squeeze(1) * right_loss + (1 - replace_flag.squeeze(1)) * self.best_loss
+        new_xs = replace_flag * right_xs.contiguous().view(b_sz,-1) + (1 - replace_flag) * new_xs
         self.i += 1
-        if self.i == 2 ** self.h or iend == dim:
-            self.h += 1
-            self.i = 0
-            # if h is exhausted, set xo_t to be xs_t
-            if self.h == np.ceil(np.log2(dim)).astype(int) + 1:
-                self.xo_t = xs_t.clone()
-                self.h = 0
-                print("new change")
-                
-        if self.p == '2':
-            import pdb
-            pdb.set_trace()
-
-        return new_xs, torch.ones(_shape[0]) + add_queries
-
-    def get_gs(self):
-        """
-        return the current estimated of the gradient sign
-        :return:
-        """
-        return self.sgn_t
+        # number of queries: add_queries (if first iteration to init best_loss) + left queries + (right queries if any)
+        num_queries = add_queries + torch.ones(b_sz) + torch.ones(b_sz) * replace_flag.squeeze(1)
+        return new_xs.view_as(xs_t), num_queries
 
     def _config(self):
         return {
             "name": self.name, 
             "p": self.p,
             "epsilon": self.epsilon,
+            "delta": self.delta,
             "lb": self.lb,
             "ub": self.ub,
             "max_extra_queries": "inf" if np.isinf(self.max_extra_queries) else self.max_extra_queries,
             "max_loss_queries": "inf" if np.isinf(self.max_loss_queries) else self.max_loss_queries,
-            "fd_eta": self.fd_eta,
             "attack_name": self.__class__.__name__
         }
-    
-'''
 
-Original License
-   
+'''
+    
 MIT License
+
 Copyright (c) 2019 Abdullah Al-Dujaili
+
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
 in the Software without restriction, including without limitation the rights
 to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 copies of the Software, and to permit persons to whom the Software is
 furnished to do so, subject to the following conditions:
+
 The above copyright notice and this permission notice shall be included in all
 copies or substantial portions of the Software.
+
 THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
