+++ attacks/score/square_attack.py
@@ -1,3 +1,6 @@
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
 '''
 
 This file is copied from the following source:
@@ -23,10 +26,6 @@ basic structure for main:
 """
 Implements Square attacks
 """
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
 import numpy as np
 from torch import Tensor as t
 import torch
