+++ requirements.txt
@@ -3,7 +3,7 @@ scipy
 matplotlib
 dill
 pandas
-torch==1.6.0
-tensorboard==2.5.0
-tensorflow-gpu==1.15.0
+#torch==1.6.0
+tensorboard
+tensorflow-gpu
 advertorch
